+++ R/plot_pred.R
@@ -10,8 +10,29 @@
 #' @return An object of class `ggplot2::ggplot`.
 #'
 #' @examples
+#' # generate a predictor matrix
 #' pred <- mice::quickpred(mice::nhanes)
+#'
+#' # plot predictor matrix for all columns
 #' plot_pred(pred)
+#'
+#' # plot predictor matrix for specific columns by supplying a character vector
+#' plot_pred(pred, c("chl", "hyp"))
+#'
+#' # plot predictor matrix for specific columns by supplying unquoted variable names
+#' plot_pred(pred, c(chl, hyp))
+#'
+#' # plot predictor matrix for specific columns by passing an object with variable names
+#' # from the environment, unquoted with `!!`
+#' my_variables <- c("chl", "hyp")
+#' plot_pred(pred, !!my_variables)
+#' # object with variable names must be unquoted with `!!`
+#' try(plot_pred(pred, my_variables))
+#'
+#' # plot predictor matrix of mids object
+#' imp <- mice::mice(mice::nhanes, print = FALSE)
+#' plot_pred(imp)
+#'
 #' @export
 plot_pred <-
   function(data,
@@ -21,6 +42,7 @@ plot_pred <-
            square = TRUE,
            rotate = FALSE) {
     verify_data(data, pred = TRUE, imp = TRUE)
+    vrb <- rlang::enexpr(vrb)
     if (mice::is.mids(data)) {
       if (!is.null(method)) {
         cli::cli_warn(
@@ -31,10 +53,14 @@ plot_pred <-
       method <- data$method
       data <- data$predictorMatrix
     }
-    p <- nrow(data)
+    vrb_matched <- match_vrb(vrb, row.names(data))
+    p <- length(vrb_matched)
     if (!is.null(method) && is.character(method)) {
       if (length(method) == 1) {
         method <- rep(method, p)
+      }
+      if (length(method > p)) {
+        method <- method[row.names(data) %in% vrb_matched]
       }
       if (length(method) == p) {
         ylabel <- "Imputation method"
@@ -48,17 +74,10 @@ plot_pred <-
       cli::cli_abort("Method should be `NULL` or a character string or vector
                      (of length 1 or `ncol(data)`).")
     }
-    vrb <- substitute(vrb)
-    if (vrb[1] == "all") {
-      vrb <- names(data)
-    } else {
-      vrb <- names(dplyr::select(as.data.frame(data), {{vrb}}))
-    }
-    vrbs <- row.names(data)
     long <- data.frame(
       vrb = 1:p,
-      prd = rep(vrbs, each = p),
-      ind = matrix(data, nrow = p * p, byrow = TRUE)
+      prd = rep(vrb_matched, each = p),
+      ind = matrix(data[vrb_matched, vrb_matched], nrow = p * p, byrow = TRUE)
     ) %>% dplyr::mutate(clr = factor(
       .data$ind,
       levels = c(-3, -2, 0, 1, 2),
@@ -81,10 +100,10 @@ plot_pred <-
                         fill = .data$clr
                       )) +
       ggplot2::geom_tile(color = "black", alpha = 0.6) +
-      ggplot2::scale_x_discrete(limits = vrbs, position = "top") +
+      ggplot2::scale_x_discrete(limits = vrb_matched, position = "top") +
       ggplot2::scale_y_reverse(
         breaks = 1:p,
-        labels = vrbs,
+        labels = vrb_matched,
         sec.axis = ggplot2::dup_axis(labels = method, name = ylabel)
       ) +
       ggplot2::scale_fill_manual(