+++ R/plot_miss.R
@@ -26,6 +26,10 @@
 #' # object with variable names must be unquoted with `!!`
 #' try(plot_miss(mice::nhanes, my_variables))
 #'
+#' # plot larger dataset
+#' plot_miss(mice::boys)
+#' plot_miss(mice::boys, ordered = TRUE)
+#'
 #' @export
 plot_miss <-
   function(data,