+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggmice (development version)
+# ggmice 0.1.1
 
 ## New features